refactor e2e moudle 4
